+++ pinned-react/src/index.js
@@ -6,6 +6,9 @@ import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <App />
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+  
 );
